Move all example code to main.cpp to make it easier to comprehend. Get rid of unnecessary QMainWindow.
