+++ demo/CMakeLists.txt
@@ -1,15 +1,12 @@
 INCLUDE_DIRECTORIES(${CMAKE_CURRENT_BINARY_DIR})
 
-SET(sizegripitemdemo_SOURCES ../SizeGripItem.cpp MainWindow.cpp main.cpp)
-SET(sizegripitemdemo_HEADERS MainWindow.h)
-SET(sizegripitemdemo_FORMS MainWindow.ui)
+SET(sizegripitemdemo_SOURCES ../SizeGripItem.cpp main.cpp)
+SET(sizegripitemdemo_HEADERS ../SizeGripItem.h)
 
 QT4_WRAP_CPP(sizegripitemdemo_HEADERS_MOC ${sizegripitemdemo_HEADERS})
-QT4_WRAP_UI(sizegripitemdemo_FORMS_HEADERS ${sizegripitemdemo_FORMS})
 
 ADD_EXECUTABLE(sizegripitemdemo
   ${sizegripitemdemo_SOURCES}
-  ${sizegripitemdemo_HEADERS_MOC}
-  ${sizegripitemdemo_FORMS_HEADERS})
+  ${sizegripitemdemo_HEADERS_MOC})
 
 TARGET_LINK_LIBRARIES(sizegripitemdemo ${QT_LIBRARIES})
